Update patch set 1
Patch Set 1:

This was a problem on the other kabylake devices until the HW was fixed.  The workaround allowed the FAFT tests to run, but they then are likely to fail for other reasons like not always powering up properly.

Could the ap-off step keep track of the fact that it didn't do the PMIC reset and then issue it when it goes to sequence?  I suppose that would lose the recovery mode state it was trying to preserve...

Patch-set: 1
CC: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>