Update patch set 1
Patch Set 1: Code-Review-2

Daisuke and I had a fairly long discussion on this today.

My concern is that providing a second way to reference GPIOs adds ambiguity when reviewing and debugging code (which we do for 10x longer than writing code).  

Removing the suffix, or providing an alias which is used in the code, removes the indication during review and compile that a signal is backwards.

If we do this only partway, in some cases, for some GPIOs, my concern is that it will not be obvious when doing 'gpioget' whether a signal is active or not.  And it will not be obvious in the code whether this is a board / signal which is really active-high, or one which just hasn't been updated in a branch to use the new terminology.  It will be easy for engineers to make mistakes when writing code or debugging boards.

So I think we should do one of three things:

1) Keep the current design pattern.

2) Change *ALL* gpio code to use gpio_is_active() / gpio_activate().

or

3) Tag GPIOs as level-type or active-type.  A GPIO which is level-type will assert if you call gpio_activate().  A GPIO which is active-type will assert if you call gpio_set_level().  Note that this will add some latency to gpio_set_level() / gpio_get_level(), and gpio_is_activated() is also inherently slower because it has to read gpio flags before returning the pin state.

Note that if we do this, we should also change the interrupt handler flags so they're GPIO_INT_ASSERTING / GPIO_INT_DEASSERTING (because GPIO_RISING / GPIO_FALLING are level).

In any case, this is a major refactoring, not something we should slip in and start using without the team having a chance to comment in a design review.

If you want to proceed, please write a design doc and share it with chromeos-ec, then if feedback is promising, we can schedule an eng review.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2