common: Add base detection support.
This commit adds a common framework for detecting a detachable base.
The framework requires two ADC pins which are for attach and detach
detection.  It implements a basic state machine and a valid detach or
attach is left for board specific code to decide.  This may simply be
comparing the read values with given thresholds.

BUG=b:69140200
BRANCH=None
TEST=Add base specific code to meowth; enable base detection, verify it
works as expected.

Change-Id: Ie01691e2114de49fd62e4dc05036d95b46b530ae
Signed-off-by: Aseda Aboagye <aaboagye@google.com>