cr50: create ap_uart state machine
Separate ap uart detection from AP detection, so we can disable the AP
uart without enabling deep sleep. If the AP is in S3 on ARM devices,
Cr50 wont be in deep sleep, but the AP UART RX signal wont be pulled up.
In this case we need cr50 ap rx to be disabled and deep sleep to be
disabled.

This change creates a state machine to handle ap uart detection. If cr50
uses the uart to detect the AP state entirely, ap_uart_is_on/off will
call ap_is_on and off as necessary.

BUG=b:35647982
BRANCH=cr50
TEST=run firmware_Cr50DeviceState on scalet and electro

Change-Id: I81336a9e232df8d44b325eef59327a1c06a80cba
Signed-off-by: Mary Ruthven <mruthven@chromium.org>