Update patch set 4
Patch Set 4: Commit-Queue+1

> Patch Set 4:
> 
> The ANX3429 schematic matches Coral, including the pull-up on DRP_EN. I don't know what the original reasoning behind this was. I agree it seems like this could be enabled later in software.

For the ANX3429, even if we didn't pull up DRN_EN, I think would we run into 
the same disconnect problem. From page 24 of datasheet 
(http://shortn/_3m8e9CgVA7): "The DRP_EN pin of ANX3429 must be configured with 
a pull-up to AVDD33 with a 4.7K ohm resistor to enable DRP mode of the ANX3429 
with CC1 and CC2 pins pulled down to GND with Rd after system power has become 
available and ANX3429 is in either standby mode or active mode. If the DRP_EN 
pin is not pulled up to AVDD33, CC1 and CC2 will be pulled up to AVDD33 with Rp 
(DFP only state) after system power becomes available and ANX3429 is in either 
standby or active mode. The charger will detect Rp present on CC1 or CC2 and 
treats this condition as a disconnection with the UFP. The charger disables 
VBUS and stops charging the system. The ANX3429 is also unable to detect if the 
charger is connected."

It sounds like if we don't pull up DRP_EN, then the TCPC is always in DFP mode 
(causing a disconnect) instead of toggling between UFP and DFP mode.

Patch-set: 4
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Label: Trybot-Ready=0