Update patch set 2
Patch Set 2:

In my experience, usually one would want to perform a warm reboot of the AP to preserve logs. But, if a warm reboot does not work and you're doing an EC reset anyways, won't that re-sequence the chipset (and by extension re-initialize DRAM)? And if that's the case, I don't see what this is fixing...

Maybe I'm missing something, but to me it seems that this change is just making `apreset` convenient instead of working as intended.

Patch-set: 2
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>