Cr50: Add brute force resistance support.
This adds some of the ground work for hardware backed brute force
resistance on Cr50. A merkel tree is used to validate encrypted
metadata used to track login attempts.

The metadata tracks counts of failed attempts along with a
timestamp of the last failed attempt along with the secrets and any
associated parameters. Instead of storing the metadata on Cr50 an
AES-CTR is used along with an HMAC to encrypt the data so it can
be stored off-chip and loaded when needed.

The Merkle tree is used to track the current state of all the
metadata to prevent replay attacks of previously exported copies.
It is a tree of hashes whose root hash is stored on Cr50, and whose
leaves are the HMACs of the encrpted metadata.

BRANCH=none
BUG=chromium:794010
TEST=cd ~/src/platform/ec && V=1 make run-pinweaver -j

Change-Id: Id10bb49d8ebc5a487dd90c6093bc0f51dadbd124
Signed-off-by: Allen Webb <allenwebb@google.com>