grunt: Rename green LED to blue
The LED is actually blue so it seems confusing to call it green. Rename
it.

Also swap the ordering of amber and blue to match the PWM. I suppose we
could change the PWM instead if that is better.

Finally, change EC LED to EC_LED_COLOR_BLUE so that the command works as
you might expect on this board.

BUG=b:71902053
BRANCH=none
TEST=make BOARD=grunt -j10

Change-Id: I48f7d88b2e53affb002a6590df86c53942951d59
Signed-off-by: Simon Glass <sjg@chromium.org>