Update patch set 4
Patch Set 4:

it's good to have prettier LEDs,
but if we are really exceeding deadlines by 170ms, which is several orders of magnitude more than what we should expect, it would be nice to fix the chip code for this issue.

That said if running it in the low priority task doesn't have the issue while the HOOK task does, there is probably a very large offenders in the hooks.

Patch-set: 4