Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> What do you think about removing the RGB assumption here? Grunt currently has 2 PWM LEDs (blue and amber). Could this be a "PWM LED" common framework that leaves the number and color settings up to the board? It looks like it is close to this already, and might allow this to be re-used by more boards in the future.

Maybe I should add this in the documentation, but it's not required to have all channels. If you look in the board specific change that follows this change you can define the channel that isn't present as RGB_LED_NO_CHANNEL (and then of course define the impossible colors with a brightness of 0).

Patch-set: 5
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>