g: handle uart interrupt storms
UART controller of the g chip does not process line break condition
properly, instead of reporting it just once it keeps reporting
received characters of zero at maximum rate, causing RX interrupt
storms.

Value of zero is never transmitted by valid sources, but when the AP
on ARM devices goes to sleep it drives the AP UART TX line low, which
causes H1 interrupt storms.

This patch detects this condition by counting how many times in a row
a zero is received, If more than three zeros are received, this
receive channel's RX interrupts are disabled and a periodic hook is
started to check if the channel is back. A channel is considered back
if the FIFO is not full, or if there is at least one non-zero
character in the FIFO.

BRANCH=cr50
BUG=b:35647982
TEST=opened/closed Bob lid 10 times in a row. Observed messages about
     UART1 being disabled/enabled. Verified that AP console is still
     available over CCD and that H1 was falling into sleep when lid
     was closed.

Change-Id: I5164bab05d3e73a0e93190c85dfd5afc2b42dd01
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>