npcx: set eSPI speed to max supported by hardware
the npcx7 can only run eSPI at (up to) 50MHz while the npcx5 can go up
to 66MHz.  so, set the max speed to whatever the hardware can actually
do.  the bit pattern for 66MHz is "reserved" on the npcx7, so let's
not even define it in the npcx7 case.

BUG=b:72838699, b:71859563
BRANCH=none
TEST="make buildall" passes; boots on meowth

Change-Id: I428caf72a41fe58008df4624c475dafadca4a0bc
Signed-off-by: Caveh Jalali <caveh@chromium.org>