bd9995x: (Test only) Disable charger noise WA and add chgop2 control
This CL disables discharge_on_ac for the case when the battery FULL
flag is set. In addition, an EC console command has been added to
control both the DC-DC frequency and the chopper mode setting in the
register CHGOP_SET2 (0xc) register.

The console command is chgop2 <chop|dcdc> <mode> <freq>

mode:
0 = off
1 = auto
2 = on

freq:
0 = 600
1 = 875
2 = 1000
3 = 1200

With just chgop2, the current settings are displayed. When 'chgop2 chop
<mode>' the 4th parma is a don't care.

When mode == auto, either chopper mode or the lower DCDC frequency is
applied whenever the chipset power state is not S0 and chopper mode is
disabled when in S0 and the DCDC frequency is always set to 1200 kHz

Change-Id: I4c54632efcaa968fc525833c72c91840347efebe
Signed-off-by: Scott Collyer <scollyer@google.com>