EFS: Give RO chance to perform vboot
When a user presses the power button in recovery screen, the system
shuts down but EC stays in RO. In this state if a user presses the
power button again, EC won't be given a chance to run EFS.

This patch makes EC execute cold reboot when a power button is pressed
in G3/S5.

BUG=b:73609782
BRANCH=none
TEST=Shut down system in recovery screen. Press power button and verify
cold boot is executed.

Change-Id: I64af0997beae707680792dcf9e656e036c2a9d8e
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>