Update patch set 3
Patch Set 3: Commit-Queue+1

> Patch Set 3: Code-Review+2
> 
> I think the spec has been this way for some time now,
> we kept it like this though after other adventures:
> on most of our other systems, switching VCONN is somewhat faster than VBUS and you might want to read b/35551927#comment12 and why DingDong doesn't like this...

That is unfortunate for DingDongs. In light of that, I tried moving the Vconn sourcing to right after vbus (before enabling TCPC RX), and it was still out of spec (~3ms instead of <2ms).

I am still going to submit it as is to ensure grunt is within spec; it just make me a little more nervous. Thanks for the background on the DingDongs though.

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1