grunt: Add power/battery LED support
Replace the current led code with an implementation that uses the new
led_pwm interface. Grunt has a blue LED which we use for power and an
amber LED which we use for battery.

The colours used are documented in update_leds() in led_pwm.c

BUG=b:71902053
BRANCH=none
TEST=manual
Check for the various states:
 * Solid Amber == Charging
 * Solid Green == Charging (near full)
 * Fast Flash Red == Charging error or battery not present

Did not test low / critical battery

Change-Id: Ie46075855ab17e6e7301025b62e57db2c596b2a4
Signed-off-by: Simon Glass <sjg@chromium.org>