tpm: ccd: allow alternative commands when ap is held in reset
A typical CCD use case is when the DUT is not fully functional, or
even completely dead, including corrupted AP/EC firmware.

We still want to be able to enable CCD in this case, but routing CCD
commands through TPM task context (necessary to ensure the large stack
size some CCD commands require) is blocked if TPM was not reset after
startup.

Let's allow both reset requests and alternative commands when AP is
held in reset after reboot. The only situation when the alternative
command arrives would be the CCD use case of the system not generating
TPM reset pulse at startup and the operator is trying to execute a CCD
or RMA reset command.

BRANCH=cr50, cr50-mp
BUG=b:73292631

TEST=on a reef device: destroyed AP firmware and observed that the
     'ccd open' command indeed results in the hung Cr50 console.

     Loaded the new Cr50 image, took it through the 'ccd open'
     sequence resulting in enabling AP flash write access, restored
     the AP flash and observed the DUT boot into Chrome OS.

Change-Id: I4413bc200f5b2be563ba666ff80dd2d889ae5790
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>