pd: retry DR_SWAP after a WAIT
If the other side answers our DR_SWAP request with a WAIT message (as
bizarre as it might be), we want to retry later rather than giving up.
Set again the PD_FLAGS_CHECK_DR_ROLE flag in order to do so.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=fizz
BUG=b:72387533
TEST=do repeated cold reset on Fizz connected to the 'interesting' USB-C
screen and see the DP alternate mode negotiated.

Change-Id: Icfb9fa209ce8cbe3c42c1e6946b43464544bdaea