Update patch set 4
Patch Set 4:

Sorry for the delay in providing this comment. I wanted to verify this before commenting.
Please note that the max possible eSPI frequency depends also on the FMCLK clock frequency. 

In NPCX5, this is defined in the datasheet in the Core Clock Frequency Settings table. In NPCX7, this is defined in the Frequency Multiplier Setting table.

NPCX7 based systems use FMCLK of 45 MHz, so eSPI frequency can be up to 50 MHz (as the max eSPI frequency supported by NPCX7). 

However, NPCX5 based systems use FMCLK of 30 MHz and in that case eSPI frequency can only be up to 33 MHz. 

Running at a higher frequency than defined may cause a timing issue.

I think the code should be updated to support max eSPI clock frequency not only per chip family but also taking into account the actual FMCLK.

Patch-set: 4