bd9995x: Use battery spec for precharge current limit
Previously, the funciton charger_set_current would always sit the
precharge current the minimum of the maximum allowed setting and what
was requested by the charge_v2 state machine. This means that the
precharge current command will regularly exceed what is specified by
the battery manufacturer for this state. Normally, this isn't an issue
because in cases where precharge state is active, the requested value
is the value specified for the battery.

However, the selection of trickle vs pre vs fast charge is not
controlled by the charge v2 state machine, but rather is controlled by
the bd9995x based on current battery and vsys values. For additional
safety, this CL limits the precharge current command to be the minimum
of the requested value and the battery's specified value plus one 64
mA step.

BRANCH=coral
BUG=b:73144529
TEST=Manual Using EC console command bd9995x_dump, verified that the
value in register 0x15 (IPRECH_SET) was set to 64 + the precharge
current setitng specified in battery.c

Change-Id: I0c6af19123cf9e2e7455a47ab38ec132418f71d7
Signed-off-by: Scott Collyer <scollyer@google.com>