Update patch set 1
Patch Set 1:

> I had to add the ABI for LLSR since we use it when handing the labels. It might make sense for me to write integration tests or reuse the unit tests to check that everything is working when it executes on actual Cr50 hardware.

It should go to or before the CL that introduces the code that handles the labels then. I like 'before that CL' better as it gives us a set of smaller, more focused CLs, which are easier to review.

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>