Fizz: Set cooling-required-temperature to 55 celsius
This patch sets 'fan_off' value of thermal_params to 55 degrees in
celsius. The CPU does not need cooling below 55 degree.

This does not completely remove 'Fan 0 stalled' because in S5/G3,
the fan is not powered but the CPU temperature can be higher than
55.

This patch also removes the two remote sensors. One doesn't exist.
The other is placed next to the local sensor. So, it's redundant.

BUG=b:73127788
BRANCH=none
TEST=Verify no 'Fan 0 stalled' is printed in S3 and S5. Verify the
fan spins in S0.

Change-Id: I8cf86c20491e95ec2b546777870584d57c700997
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>