gsctool: add open box RMA option
This patch enhances the gsctool utility to allow to verify RO sections
of the target AP and EC flash memory.

The only command line parameter required for the new option ('O') is
the file name of the target descriptors database, containing memory
description sections for one or more Chrome OS devices.

Memory description sections are of two types (both types could be
referring AP or EC memory):

 - hash descriptor, this section includes the address range of the
   memory and one or more hash values for the contents of that address
   range. Multiple hashes are needed in case when the same device has
   mnore than one RO firmware releases in circulation.

 - dump descriptor, this is a request for this utility to display on
   the console the contents of the certain area of flash memory on the
   target.

When this utility starts the process, the target might request that
the operator confirms physical presence, in this case the utility
keeps prompting the operator to press the physical presence button
until DUT is satisfied,

BRANCH=none
BUG=b:73668125
TEST=created a descriptor database for a Robo device feeding it with
     values retrieved on the device by locally running spihash command
     on the device.

     Then ran this utility to verify successful hash and dump
     retrievals, comparing dump values with values obtained through
     Cr50 console directly.

     Created additional dummy hash variants and verified that the
     utility succeeds only if all matches happen at the same variant
     index in different hash sections.

Change-Id: Ib43cf4eb642d141b7cd7f129ef412e14bd59f30b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/933545
Reviewed-by: Mary Ruthven <mruthven@chromium.org>