pd: Don't auto toggle when DRP state is force sink
If we are in PD_STATE_DRP_AUTO_TOGGLE and DRP state is PD_DRP_FORCE_SINK,
then we remain in PD_STATE_DRP_AUTO_TOGGLE when a sink is connected.

A USB-C hub with pass-through charging will be toggling between source
and sink so might be seen as a sink depending on the timing of the
connection.

DRP state is set to PD_DRP_FORCE_SINK when we shutdown (S5/G3).

This means that if a charger is attached via a hub when we are shutdown,
sometimes we will not see the hub as a source, even though our DRP state
is PD_DRP_FORCE_SINK. This brokenness persists until the hub is unplugged.

To avoid this, stay in state PD_STATE_SNK_DISCONNECTED when DRP state is
PD_DRP_FORCE_SINK, instead of moving to PD_STATE_DRP_AUTO_TOGGLE.

BRANCH=none
BUG=b:72007056
TEST=connect charger via hub to Grunt in S5 multiple times

Change-Id: Idcf4a57cc73ecf375b1f430712a8cfce892ada80
Signed-off-by: Edward Hill <ecgh@chromium.org>