npcx/fan: Make all fans in RPM mode work
Currently, if there is a fan which isn't in RPM mode, the rest of the
fans are not controlled even if they're in RPM mode. This patch gives
a chance to all the fans.

BUG=b:35543471
BRANCH=none
TEST=none

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/940813
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit 84208c27ae14003e980575df7b31b9c2aab1c40e)

Change-Id: Ic15318b8200f35a3868aa6d3403916249b69fb55