cr50: manage RMA key inclusion
We want to be able to test RMA unlock functionality in dev builds
without requiring the communicating with the server involving u2f.

We also want to make sure that images signed for production do not
by accident incorporate test RMA keys.

The suggested by this patch way of handling the issue is as follows:

 - by default the prod RMA server public key is incorporated in the
   image. If USE_TEST_RMA_KEY is defined as an non empty string in the
   environment or as a variable in the make invocation, the test RMA
   server public key is included in the image.

 - the script used to sign prod Cr50 image scans both elf files passed
   in to check if test RMA key pattern can be found there, and if so -
   the script would refuse to sign the image.

BRANCH=cr50, cr50-mp
BUG=b:70891959
TEST=verified that invoking make as

    $ USE_TEST_RMA_KEY=1 make BOARD=cr50

    triggers inclusion of the test RMA key in the image.

    Also verified that create_relleased_image.sh rejects images
    including test key.

Change-Id: I2f5934e3052cdf9bc4c9c1db98d0a16e6c5983f9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>